Typecast council dist number for merge
Former-commit-id: d4efdfbc70ad31a007091f2e1f8bdb5411a6a5be

+++ developments_build/sql/aggregate.sql
@@ -583,6 +583,6 @@ SELECT
 INTO AGGREGATE_councildst
 FROM _AGGREGATE_councildst a
 JOIN council_members b
-ON a.councildst = b.district
+ON a.councildst::int = b.district::int
 GROUP BY a.councildst, b.name
 ORDER BY councildst;
